Addition of `primary_sites_counts.tsv`
Edited the `01-pbta-analysis-5.R` script to produce a TSV file that has each unique primary_site, the total number of tumor types expressed at that site, along with the highest and second highest expressed tumor types at each site.

The `primary_sites_counts.tsv`file in `analyses/sample-distribution-analysis/results" is intended to be a convenient summary of the data to be put on the brain graphic.