Addition of README, install packages, and initial analysis scripts
Additions:
`00-README.md` outlines the contents and structure of the sample-distribution-analysis folder.
`00-install-packages.R` installs packages required for the execution of scripts to follow.
`01-pbta-analysis-5.R` performs the initial filtering of the dataset. It produces a PDF and TSV of the sample distribution across cancer types, and TSV files of the cancer type distribution across primary sites of the brain. 
`.gitignore` was edited to include the additional files produced when HTML outputs are made (these outputs are made in an upcoming script).