Fix umap_plot
Using the same argument name as one of the columns was causing the `comparable` argument to be ignored, so the name and behavior of the argument was changed.

Also includes standardizing on poly-A over polyA in text and formatting changes, and using the roughly filtered data for the raw clustering.