Check unchanged files in previous release and don't download again
If the previous release folder exists, check the MD5 of files in that folder against the new md5 manifest. If the old files match, hard link them into the new folder. Then proceed to download any remaining files not present.

Also, don't redownload reference files if present. (They should probably be checked as well, but are not currently in the manifest.)

Final md5 check is still performed for all files.