Reuse previously downloaded files in download-data.sh (#301)
* :wrench: update download-data.sh for v11 release

* :wrench: update release-notes.md for v11 release

* update release-notes; replace consensus zip

* change consensus MAF to gzipped

* Remove unzip step

* Check unchanged files in previous release and don't download again

If the previous release folder exists, check the MD5 of files in that folder against the new md5 manifest. If the old files match, hard link them into the new folder. Then proceed to download any remaining files not present.

Also, don't redownload reference files if present. (They should probably be checked as well, but are not currently in the manifest.)

Final md5 check is still performed for all files.

* Don't hard link over existing files

Also, check symlinks at the end.

* Hide a warning

* Remove --ignore-missing

It doesn't seem to be necessary, as we will get a non-zero return code anyway if any files changed.

* Move old symlink deletion back to start