Handle missing data, and some refactoring
I made some substantial changes here in structure, but the results should be largely unchanged.

I did some transposing when constructing data frames so we can use the same function (fill_df)  to extract data more often, and moved the ID column specification out of the function so that RNA and DNA-derived data are handled the same way.

The function then allows a set of samples to be specified, and if the request is for a a sample that does not fall in there, it is set as NA for that column in the output data, otherwise it is filled in with a default value.