Select stably expressed genes
Select genes that are stably expressed in poly-A and stranded RNA-seq libraries
by the following criteria, and the filtered output
`results/stranded_vs_polya_stably_exp_genes.csv` is ordered by the coefficient
of variations.

   - normalized read count coefficient of variation across all biospecimens < 0.5
   - `exactTest`, `glmLRT`, and `glmQLFTest` p-values all > 0.05
   - logFC < 0.5
   - logCPM > 0.8

Generate boxplots for comparing the poly-A and stranded normalized read counts
of the top stably expressed genes.

Generate logFC and coefficient of variation histograms for selecting stably
expressed genes.