Change default to add all annotation columns
The interface of annotate_long_format_table is changed, so that all
annotation columns are added by default, and the columns_to_add
parameter can be passed to add only certain columns.

The current interface annotate_long_format_table is designed to favor
flexibility and readability over efficiency. The funciton can be further
refactored to be more efficient, e.g. not check certain tables, process
certain tables, or add certain columns if they are not required.