Specify annotation data versions in README.md
Add annotation data versions and data of the last update in the "Update
downloaded data that are used in this module" section, as suggested by
@jharenza at
https://github.com/PediatricOpenTargets/OpenPedCan-analysis/pull/55#discussion_r673386812

Combine gene and disease (/cancer_group) annotations into one table.

Add additional notes on annotation data versions to the "Implementation
of long-format table annotator" section.