Specify envir = parent.frame() in eval call
In import_function, envir = parent.frame() makes the
environment(imported_function) to be the same as the import function
being called.

Even though the eval documentation says the default envir parameter is
parent.frame(), leaving envir = parent.frame() in the eval call will
surprisingly make the environment(imported_function) to be the same as
the environment of the import_function call. The tests also fail without
specifying envir = parent.frame(). Maybe this is caused by the place
where parent.frame() is evaluated? If specified, parent.frame() is
evaluated in the calling function; if not specified, parent.frame() is
evaluated in the eval call environment?

Examples executed in terminal R, in order to avoid RStudio customizations.
> print(environment())
<environment: R_GlobalEnv>
>
> foo <- function() {
+   print(parent.frame())
+   print(environment())
+   return(eval(quote(function() { return(2) })))
+ }
>
> bar <- foo()
<environment: R_GlobalEnv>
<environment: 0x5645ead67670>
> print(environment(bar))
<environment: 0x5645ead67670>
>
> baz <- function() {
+   print(parent.frame())
+   print(environment())
+   return(eval(quote(function() { return(2) }),
+               envir = parent.frame()))
+ }
>
> qux <- baz()
<environment: R_GlobalEnv>
<environment: 0x5645ead5fec0>
> print(environment(qux))
<environment: R_GlobalEnv>