Test annotator CLI --replace-na-with-empty-string
Test that --replace-na-with-empty-string replaces NA with "".

Test that no --replace-na-with-empty-string does not replace NA with "".

The --replace-na-with-empty-string does not need to be tested for column
type conversions. The CLI write_tsv output table is the same even if a
non-character column is converted to a character column before
write_tsv, because "Values are only quoted if they contain a comma,
quote or newline" (-- help("write_tsv", "readr") 1.3.1).