Remove MAF rows that have NA in variant grouping columns
Remove rows that have NA in columns that are used as the variant
grouping column when computing frequencies using get_opr_mut_freq_tbl.

The results are identical as before, because no row is removed by this
additional filtering step, but this step prevents future NAs being
introduced to variant IDs or gene ENSG IDs that are used to compute
mutation frequencies.