Make result xlsx file identically reproducible
Previously, the md5sum of result xlsx file is different in each run of
01-generate-pedot-column-display-order-name-xlsx.py, even though the
content of the xlsx file is identical to each other. The difference is
caused by the creation time of the xlsx file.

An xlsx file has the following two places for creation date time:

- openpyxl workbook properties
- ZipInfo. xlsx file is actually a zip file,
https://foss.heptapod.net/openpyxl/openpyxl/-/blob/branch/3.0/openpyxl/writer/excel.py#L291-293
The zip file has ZipInfo to record current local time
https://github.com/python/cpython/blob/282fc5c00787297f1d7a9b7b1469b51e84f485a9/Lib/zipfile.py#L1764-L1775
.

Now, the result xlsx file is identically reproducible. The xlsx file is
created using previous method with property creation date time set.
Then, the date time in ZipInfo is updated with
utils.update_xlsx_datetime .